bind text-area-field
