+++ packages/js/product-editor/src/bindings/extend-blocks-with-bound-attributes/block-edit-with-binding-attrs.tsx
@@ -13,7 +13,7 @@ import {
 /**
  * Internal dependencies
  */
-import { BLOCK_BINDINGS_ALLOWED_BLOCKS, isBlockAllowed } from './';
+import { BLOCK_BINDINGS_ALLOWED_BLOCKS, isBlockAllowed } from '../';
 import productEntitySource, {
 	type CoreBlockEditProps,
 } from '../product-entity-source/index';
