+++ packages/js/product-editor/src/bindings/extend-blocks-with-bound-attributes/index.ts
@@ -6,23 +6,7 @@ import type { BlockConfiguration } from '@wordpress/blocks';
  * Internal dependencies
  */
 import blockEditWithBoundAttribute from './block-edit-with-binding-attrs';
-
-type BLOCK_BINDINGS_ALLOWED_BLOCKS_TYPE = {
-	[ key: string ]: string[];
-};
-
-export const BLOCK_BINDINGS_ALLOWED_BLOCKS: BLOCK_BINDINGS_ALLOWED_BLOCKS_TYPE =
-	{
-		'core/paragraph': [ 'content' ],
-		'core/heading': [ 'content' ],
-		'core/image': [ 'url', 'title', 'alt' ],
-		'core/button': [ 'url', 'text', 'linkTarget' ],
-		'woocommerce/product-text-area-field': [ 'content', 'placeholder' ],
-	};
-
-export function isBlockAllowed( blockName: string ): boolean {
-	return blockName in BLOCK_BINDINGS_ALLOWED_BLOCKS;
-}
+import { isBlockAllowed } from '..';
 
 export default function extendBlockWithBoundAttributes(
 	settings: BlockConfiguration,
