+++ packages/js/product-editor/src/bindings/index.ts
@@ -10,6 +10,24 @@ import { addFilter } from '@wordpress/hooks';
 import productEntitySource from './product-entity-source';
 import extendBlockWithBoundAttributes from './extend-blocks-with-bound-attributes';
 
+type BLOCK_BINDINGS_ALLOWED_BLOCKS_TYPE = {
+	[ key: string ]: string[];
+};
+
+export const BLOCK_BINDINGS_ALLOWED_BLOCKS: BLOCK_BINDINGS_ALLOWED_BLOCKS_TYPE =
+	{
+		'core/paragraph': [ 'content' ],
+		'core/heading': [ 'content' ],
+		'core/image': [ 'url', 'title', 'alt' ],
+		'core/button': [ 'url', 'text', 'linkTarget' ],
+		'woocommerce/product-text-area-field': [ 'content', 'placeholder' ],
+		'woocommerce/product-name-field': [ 'name' ],
+	};
+
+export function isBlockAllowed( blockName: string ): boolean {
+	return blockName in BLOCK_BINDINGS_ALLOWED_BLOCKS;
+}
+
 const registerBlockBindingsSource =
 	// @ts-expect-error There are no types for this.
 	dispatch( blockEditorStore )?.registerBlockBindingsSource;
@@ -20,7 +38,7 @@ const registerBlockBindingsSource =
  *
  * @return {boolean} Whether the block binding API is available.
  */
-export function isBlockBindingAPIAvailable() {
+export function isBlockBindingAPIAvailable(): boolean {
 	const isAvailable = Boolean( registerBlockBindingsSource );
 	if ( ! isAvailable ) {
 		console.warn( 'Binding API not available' ); // eslint-disable-line no-console
